add gpt scripts
